backend/pyproject.toml
[project]
name = "scrap_llm"
version = "0.1.0"
description = ""
authors = [
    { name = "Younes Belataris", email = "younes.belataris@appchoose.io" },
    { name = "Chadlia El Amri", email = "chadlia.elamri@appchoose.io" },
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "uvicorn (>=0.34.3,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "requests (>=2.32.3,<3.0.0)",
    "crawl4ai (>=0.6.3,<0.7.0)",
    "furl (>=2.1.4,<3.0.0)",
    "langchain-openai (>=0.3.19,<0.4.0)",
    "langfuse (>=2.60.7,<3.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "langchain-anthropic (>=0.3.14,<0.4.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "google-adk (>=1.1.1,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.33.1,<2.0.0)",
    "PyJWT (>=2.10.1,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "google-cloud-tasks (>=2.19.2,<3.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "plotly (>=6.1.2,<7.0.0)",
    "regex (>=2024.11.6,<2025.0.0)",
    "sentry-sdk[fastapi] (>=2.30.0,<3.0.0)",
]

[tool.poetry]
packages = [
    { include = "scrap_llm", from = "src" },
    { include = "benchmarks", from = "." },
]

[tool.poetry.scripts]
serve = "app.main:run"
benchmarks = "benchmarks.app:launch"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
